files/requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
python-dotenv==1.0.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
asyncio==3.4.3
colorama==0.4.6
